+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#23](https://github.com/LaxarJS/ax-input-control/issues/23): users can now `$broadcast` the event `axInput.setPristine` to reset the validation state
+    + NEW FEATURE: see ticket for details
 
 ## v2.0.0-alpha.7